I think we may have the pre-processing script working!
I think it works! I think we may have too many quote marks around the UUIDs, and we need to confirm that we have all the backslashing right so that the programs in the DB are still running, but we're close if not there. We also still need to turn all the test case errors into a single array, and convert all the appropriate strings to numeric values when we do the import.