Figured out an alternative to OR for combining sets.
Using Tinkerpop's `or` or `union` commands to combine two sets/streams
(the individuals with total error zero and the individuals from the
last generation) was unusably slow. This extracts those two sets of
individuals separately (which is fast), stores them in groovy lists,
and then uses `inject` and `unfold` to build a new pipeline out of those
vertices.

I have not idea if this is the "best" way to do this, but it does work.