A first pass at modularizing `LoadCsv.groovy`.
This code was quite a mess, and this is a first pass at refactoring it into
some useful functions. I needed to convert two functions from defs to closures
because otherwise they couldn't seem to "see" each for reasons I don't really
understand. I think it has something to do with the implicit class that
the `groovysh` uses to wrap code we `:load`, but I'm not at all clear on the
details.

Several of these functions are still *way* longer than I'd like, and there's
still a bit more fiddling after the function definitions than is optimal, but
it's a significant step in a useful direction.