Created the "master function" for loading CSV files.
The `loadCsv` function takes the name of the properties file (which includes
the location of the DB files) and the name of the CSV file, and loads the
given CSV file into the DB specified in the properties files.

This makes all of the contents of `LoadCsv.groovy` functions now, with nothing
that actually runs any code. This frees us to load up these definitions, and
then call them without having to edit/hack at the code. (That's not
*completely* true, sadly, because of weird dependencies on column names and
numbers that change from problem to problem, but we're getting there.)