Added computation of Damerau-Levenshtein distance on edges.
This extracts a list of the instructions and close counts for
every node, and computes the D-L distance between those two
arrays, storing that value on the associated edge in the graph.

This definitely slows down the DB build process (maybe a factor
of 2?), but it's a one-time cost and really useful, so that's not
a huge issue. We do, at the moment, generate the array of items two
or three times, once when that individual is the parent, and once
for every time it's a child. I don't know how expensive that is, but
if it's significant then avoiding that would speed things up. I'm
not sure that's the main source of the slow down, though.