Add Clojure/Neocons scripts for Neo4j
I've been working on a set of Clojure scripts using the
Neocons library for talking to Neo4j DBs. This is where
this sits at the moment, with a number of functions that
augment a DB in useful ways (e.g., add "ancestorOfWinner"
fields) and compute Page Rank using the APOC library.

One big question is how to best use Clojure's parallelism
tools. If I do some "obvious" things with `pmap` it sometimes
works, but it sometimes leads to deadlock problems in Neo4j
that I don't fully understand.