Parallelized Page Rank commits
When I tried running the APOC Page Rank implementation on the
genes in a significant run, the Page Rank calculation didn't take
_too_ long (under an hour?), but committing all the new property
values to all those nodes ran for over a day without any sign of
completion. I modified the Clojure code for this so that it receives
the results of the APOC Page Rank calculation and then commits each
node's Page Rank value as a separate query in parallel.

This is definitely making progress (over 6 million commits in about
an hour), but still might be too slow to be very usable. I suspect
it's not super efficient to every *every* node be a separate commit,
but I'm not sure how to avoid that.

I think it's still going to be way too slow to be useful. That said,
we don't really need 100 million nodes to have Page Rank values. It
would probably be sufficient to do something like limit the return to
the 1 million with the highest Page Rank. That would probably complete
in a couple of hours, and I doubt we really need more than that?